% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_tags}
\alias{get_tags}
\title{Get tags associated with a Term}
\usage{
get_tags(term, ...)
}
\arguments{
\item{term}{the term to search for.}

\item{...}{further arguments to pass to httr's GET function.}
}
\value{
a data.frame containing the term and the top 10 tags associated with it.
}
\description{
\code{get_tags} retrieves the tags associated
with a particular term on Urban Dictionary.
}
\examples{
# Simple example - retrieve the entry for "dab"
fade_definitions <- get_tags("dab")

}
\seealso{
\code{\link{get_term}} for retrieving a set of \emph{definitions} for
a specific term.
}

