% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_term}
\alias{get_term}
\title{Retrieve Urban Dictionary definitions for a term}
\usage{
get_term(term, ...)
}
\arguments{
\item{term}{the term to search for.}

\item{...}{further arguments to pass to httr's GET function.}
}
\description{
\code{get_term} gets definitions for a specific term from
Urban Dictionary, along with associated metadata such as the submitter,
up- and down-votes, and a permalink.
}
\examples{
# Simple example - retrieve the entry for "friendly fade"
fade_definitions <- get_term("friendly fade")

}
\seealso{
\code{\link{random_term}} for retrieving a set of random definitions
for random terms.
}
