% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{random_term}
\alias{random_term}
\title{Retrieve random Urban Dictionary definitions}
\usage{
random_term(...)
}
\arguments{
\item{...}{arguments to pass to httr's GET function.}
}
\description{
\code{random_term} retrieves a random set of 10
Urban Dictionary definitions, along with associated metadata.
}
\examples{
# Grab 10 random entries
entries <- random_term()

}
\seealso{
\code{\link{get_term}} for retrieving a set of definitions for
a specific term.
}
