% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udapi.R
\docType{package}
\name{udapi}
\alias{udapi}
\alias{udapi-package}
\title{Urban Dictionary API client}
\description{
a client library for Urban Dictionary's API
}
\seealso{
\code{\link{get_term}} for retrieving definitions for a specific term,
\code{\link{random_term}} for random definitions of random terms, and \code{\link{get_tags}} for
tags associated with a term.
}

